mobile/tests/e2e/requirements.txt
pytest>=8.0
pytest-playwright>=0.5
httpx>=0.28
